move code table
